+++ ca/ca/test_settings.py
@@ -104,6 +104,7 @@ INSTALLED_APPS = (
     'django.contrib.admin',
     # Uncomment the next line to enable admin documentation:
     # 'django.contrib.admindocs',
+    'django_object_actions',
 
     'django_ca',
 )