+++ docs/source/changelog.rst
@@ -10,11 +10,13 @@ ChangeLog
 1.10.0 (TBR)
 ************
 
+* New dependency: `django-object-actions <https://github.com/crccheck/django-object-actions>`_.
 * Add link to easily download certificate bundles.
 * Support ECC private keys for new Certificate Authorities.
 * Store CA private keys in the more secure `PKCS8 format
   <https://cryptography.io/en/latest/hazmat/primitives/asymmetric/serialization/#cryptography.hazmat.primitives.serialization.PrivateFormat.PKCS8>`_.
 * Run test-suite with all warnings comming from **django-ca** turned into errors.
+* The Certificate change view now has a second "Revoke" button as object action next to the "History" button.
 * Also test with Python 3.7.0.
 
 .. _changelog-1.9.0: